Make xunit.yml workflow more stable and able to accept missing coverage reports (#5)
Primary issue was that when `grep` doesn't find any matches, it throws an error and the workflow dies. Using `grep [pattern] || true` seems to fix this issue.

Also added more verbose debugging output, which is only included when the calling workflow sets `DEBUG` to `true`.

Improved logging & formatting for the first-time reporting case.