Fix shared configs workflow steps
The condition on the dotnet.yml workflow that copies in JSON configuration files was faulty.
The `CONFIGS` input parameter should contain the name of the repo to copy configs in from, however the `if` statement on the workflow step assumed it was a boolean. The condition should be true if the input is non-empty.